test case 2: completed test case 3: started
